requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Scientific computing
scikit-learn>=1.0.0
statsmodels>=0.13.0

# Signal processing & EDF
mne>=1.0.0  # For EDF file reading and sleep staging
pyedflib>=0.1.30  # Alternative EDF reader

# Bayesian inference & HMM
# pyhsmm>=0.1.8  # Unmaintained but usable for HDP-HMM experiments
# Note: pyhsmm may require manual installation:
# pip install git+https://github.com/mattjj/pyhsmm.git

# Alternative: ssm for modern HMM implementations
ssm>=0.0.1  # State-space models

# Plotting
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0  # For interactive visualizations

# Data handling
h5py>=3.1.0  # For efficient storage of features
tables>=3.6.0  # PyTables for HDF5

# Configuration
pyyaml>=5.4.0
omegaconf>=2.1.0  # Hierarchical configs

# Utilities
tqdm>=4.62.0  # Progress bars
joblib>=1.0.0  # Parallel processing
click>=8.0.0  # CLI argument parsing

# Jupyter & notebooks
jupyter>=1.0.0
ipython>=7.25.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Evaluation metrics
scikit-learn>=1.0.0  # Already listed, includes ARI, NMI
munkres>=1.1.4  # Hungarian algorithm for label alignment

# MCMC diagnostics
arviz>=0.11.0  # Bayesian inference diagnostics (R-hat, ESS, trace plots)

# Optional: JAX for modern implementations
# jax>=0.3.0
# jaxlib>=0.3.0
# tensorflow-probability>=0.16.0  # For TFP distributions in JAX

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0

# Code quality
black>=21.6b0  # Code formatter
flake8>=3.9.0  # Linter
mypy>=0.910  # Type checking

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0
